+++ app_types/dorm_types.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class DormGender(str, Enum):
+    MAN = "גברים"
+    WOMAN = "נשים"
+    MIXED = "מעורב לא עלינו רחמנא ליצלן"
+    SPECIALS = "להטבק"
