+++ models/dorms.py
@@ -1,13 +1,15 @@
+import datetime
 from typing import List, Optional
 
 from sqlmodel import Field, SQLModel, Relationship
+from uuid6 import UUID,uuid7
 
-from dormtypes.soldier_types import DormGender
+from app_types.dorm_types import DormGender
 from models.users import User
 
 
 class Room(SQLModel, table=True):
-    id: int = Field(primary_key=True)
+    id: UUID = Field(primary_key=True, default_factory=uuid6.uuid7,index=True)
     name: Optional[str]
     capacity: int = Field(ge=1)
     soldiers: List[User] = Relationship(back_populates="room")
@@ -16,7 +18,8 @@ class Room(SQLModel, table=True):
 
 
 class Dormitory(SQLModel, table=True):
-    id: int = Field(primary_key=True)
+    id: UUID = Field(primary_key=True, default_factory=uuid7, index=True)
+    name: str
     gender: DormGender
     rooms: List[Room] = Relationship(back_populates="dorm")
     base_id: int = Field(foreign_key=True)
@@ -24,5 +27,6 @@ class Dormitory(SQLModel, table=True):
 
 
 class Base(SQLModel, table=True):
-    id: int = Field(primary_key=True)
+    id: UUID = Field(primary_key=True, default_factory=uuid7, index=True)
+    name: str
     dorms: List[Dormitory] = Relationship(back_populates="base")
