+++ pyproject.toml
@@ -13,7 +13,8 @@ dependencies = [
     "pydantic-settings (>=2.13.1,<3.0.0)",
     "sqlalchemy (>=2.0.47,<3.0.0)",
     "alembic (>=1.18.4,<2.0.0)",
-    "sqlmodel (>=0.0.37,<0.0.38)"
+    "sqlmodel (>=0.0.37,<0.0.38)",
+    "uuid6 (>=2025.0.1,<2026.0.0)"
 ]
